Move common logic out of fio/dlio parse_logs.py
