+++ perfmetrics/scripts/testing_on_gke/examples/fio/parse_logs.py
@@ -26,7 +26,8 @@ import sys
 # local library imports
 sys.path.append("../")
 import fio_workload
-from utils.utils import get_memory, get_cpu, unix_to_timestamp, is_mash_installed, download_gcs_objects
+from utils.utils import get_memory, get_cpu, unix_to_timestamp, is_mash_installed
+from utils.parse_logs_common import ensureDir, download_gcs_objects, parseLogParserArguments
 
 _LOCAL_LOGS_LOCATION = "../../bin/fio-logs"
 
@@ -49,13 +50,6 @@ record = {
     "filesPerThread": 0,
     "numThreads": 0,
 }
-
-
-def ensureDir(dirpath: str):
-  try:
-    os.makedirs(dirpath)
-  except FileExistsError:
-    pass
 
 
 def downloadFioOutputs(fioWorkloads: set, instanceId: str) -> int:
@@ -86,48 +80,6 @@ def downloadFioOutputs(fioWorkloads: set, instanceId: str) -> int:
       print(f"Failed to download FIO outputs from {srcObjects}: {errorStr}")
       return returncode
   return 0
-
-
-def parseLogParserArguments() -> object:
-  parser = argparse.ArgumentParser(
-      prog="DLIO Unet3d test output parser",
-      description=(
-          "This program takes in a json workload configuration file and parses"
-          " it for valid FIO workloads and the locations of their test outputs"
-          " on GCS. It downloads each such output object locally to"
-          " {_LOCAL_LOGS_LOCATION} and parses them for FIO test runs, and then"
-          " dumps their output metrics into a CSV report file."
-      ),
-  )
-  parser.add_argument(
-      "--workload-config",
-      help=(
-          "A json configuration file to define workloads that were run to"
-          " generate the outputs that should be parsed."
-      ),
-      required=True,
-  )
-  parser.add_argument(
-      "--project-number",
-      help=(
-          "project-number (e.g. 93817472919) is needed to fetch the cpu/memory"
-          " utilization data from GCP."
-      ),
-      required=True,
-  )
-  parser.add_argument(
-      "--instance-id",
-      help="unique string ID for current test-run",
-      required=True,
-  )
-  parser.add_argument(
-      "-o",
-      "--output-file",
-      metavar="Output file (CSV) path",
-      help="File path of the output metrics (in CSV format)",
-      default="output.csv",
-  )
-  return parser.parse_args()
 
 
 if __name__ == "__main__":
